+++ lib/flycal_cli.rb
@@ -10,6 +10,7 @@ require "flycal_cli/clipboard"
 require "flycal_cli/calendar_service"
 require "flycal_cli/slot_finder"
 require "flycal_cli/slot_formatter"
+require "flycal_cli/mock"
 require "flycal_cli/pipeline"
 require "flycal_cli/cli"