+++ lib/flycal_cli/mock.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+require "flycal_cli/mock/config"
+require "flycal_cli/mock/event_generator"
+require "flycal_cli/mock/calendar_service"
+
+module FlycalCli
+  module Mock
+  end
+end